mydotsページからbaseページへ移行
