+++ src/components/pages/Base.jsx
@@ -1,20 +1,168 @@
-import React, { useEffect, useContext } from "react";
+import React, { useEffect, useContext, useState } from "react";
+import Header from "../templates/Header/Header";
+import Footer from "../templates/Footer/Footer";
+import { makeStyles } from "@material-ui/core/styles";
 import { useDispatch } from "react-redux";
 import { Link } from "react-router-dom";
-import Footer from "../templates/Footer/Footer";
-import Header from "../templates/Header/Header";
-import BarChart from "../templates/graph/BarChart";
-import MiniDots from "../templates/MiniDots";
-import MiniForm from "../templates/MiniForm";
 import { AuthContext } from "../../firebase/AuthService";
-import firebase from "../../firebase/firebase";
 import { fetch_dots } from "../../reducks/dots/action";
 import { fetch_todayDotLength } from "../../reducks/star/action";
-import { Bodyleft } from "../templates/Header/HeaderElements";
+import firebase from "../../firebase/firebase";
+import MydotsChart from "../templates/graph/MydotsChart";
+import UserIcon from "../templates/icons/user/user";
+import OurSideBar from "../templates/OurSideBar";
+import Calendar from "../templates/Calendar";
+import MiniDots from "../templates/MiniDots";
+import MiniForm from "../templates/MiniForm";
+import {
+  LeftItem,
+  Profile,
+  WeekStudyHours,
+  StudyHours,
+  StyledCalendar,
+  ExplainCa,
+  StyledChart,
+  StyledForm,
+  StyledDots,
+} from "../../style/BaseStyle";
+
+const useStyles = makeStyles({
+  container: {
+    width: "400px",
+    margin: "0 auto",
+  },
+  input: {
+    width: "343px",
+  },
+  div: {
+    height: "40px",
+  },
+  text: {
+    width: "400px",
+  },
+});
+
+const MainStyle = {
+  display: "flex",
+  justifyContent: "space-between",
+};
 
 export default function Base() {
   const dispatch = useDispatch();
   const user = useContext(AuthContext);
+  const classes = useStyles();
+  const [week_hours, setWeek_hours] = useState("");
+  const [lastweek_hours, setLastweek_hours] = useState("");
+  const [total_hours, setTotal_hours] = useState("");
+  const db = firebase.firestore().collection("dots");
+
+  // 今日から一週間前の指定
+  const specify_weekago = () => {
+    let agoDate = new Date();
+    let agoWeek = agoDate.setDate(agoDate.getDate() - 6);
+    let hope = new Date(agoWeek);
+    let zero = hope.setHours(0);
+    let one = new Date(zero);
+    let two = one.setMinutes(0);
+    let three = new Date(two);
+    let four = three.setSeconds(0);
+    let five = new Date(four);
+    return five;
+  };
+
+  let weekago_yesterday = new Date(
+    specify_weekago().getFullYear(),
+    specify_weekago().getMonth(),
+    specify_weekago().getDate() - 1
+  );
+
+  const just_weekago = new Date(weekago_yesterday.setHours(23, 59, 59, 999));
+
+  //前週を指定
+  const specify_lastweek = () => {
+    let agoWeek2 = weekago_yesterday.setDate(weekago_yesterday.getDate() - 6);
+    let hope2 = new Date(agoWeek2);
+    let zero2 = hope2.setHours(0);
+    let one2 = new Date(zero2);
+    let two2 = one2.setMinutes(0);
+    let three2 = new Date(two2);
+    let four2 = three2.setSeconds(0);
+    let five2 = new Date(four2);
+    return five2;
+  };
+
+  //一週間文のdotsを取得し勉強時間を合計する
+  useEffect(() => {
+    if (user) {
+      const week_array = [];
+      db.where("userId", "==", user.uid)
+        .where(
+          "createdAt",
+          ">",
+          firebase.firestore.Timestamp.fromDate(specify_weekago())
+        )
+        .get()
+        .then((data) => {
+          data.docs.map((doc) => {
+            const item = doc.data();
+            week_array.push(item);
+          });
+          const totalWeekHours = week_array.reduce((result, current) => {
+            return result + current.working;
+          }, 0);
+          setWeek_hours(totalWeekHours);
+        });
+    }
+  }, [user]);
+
+  //前週のdotsをDBから取得
+  useEffect(() => {
+    const lastweek_array = [];
+    if (user) {
+      //firestoreのDBでも時間が指定できるように変換する
+      const startDate = firebase.firestore.Timestamp.fromDate(
+        specify_lastweek()
+      );
+      const endDate = firebase.firestore.Timestamp.fromDate(just_weekago);
+      db.orderBy("createdAt")
+        .where("userId", "==", user.uid)
+        .startAt(startDate)
+        .endBefore(endDate)
+        .get()
+        .then((data) => {
+          data.docs.map((doc) => {
+            const item = doc.data();
+            lastweek_array.push(item);
+          });
+          const total_lastweekHours = lastweek_array.reduce(
+            (result, current) => {
+              return result + current.working;
+            },
+            0
+          );
+          setLastweek_hours(total_lastweekHours);
+        });
+    }
+  }, [user]);
+
+  //総学習時間を取得
+  useEffect(() => {
+    const total_array = [];
+    if (user) {
+      db.where("userId", "==", user.uid)
+        .get()
+        .then((data) => {
+          data.docs.map((doc) => {
+            const item = doc.data();
+            total_array.push(item);
+          });
+          const total_hours = total_array.reduce((result, current) => {
+            return result + current.working;
+          }, 0);
+          setTotal_hours(total_hours);
+        });
+    }
+  }, [user]);
 
   const get_todayMidnight = () => {
     const TODAY_MIDNIGHT = new Date();
@@ -82,52 +230,33 @@ export default function Base() {
   return (
     <React.Fragment>
       <Header />
-      {/* <BarChart /> */}
-      {/* <Bodyleft> */}
-      {/* <MiniForm /> */}
-      {/* <MiniDots /> */}
-      {/* </Bodyleft> */}
+      <form className="MainFrom" style={MainStyle}>
+        <div>
+          <OurSideBar />
+        </div>
+        <LeftItem>
+          <UserIcon />
+          <StyledDots>
+            <MiniDots />
+          </StyledDots>
+        </LeftItem>
+        <StyledChart>
+          <StyledForm>
+            <MiniForm />
+          </StyledForm>
+          <MydotsChart />
+        </StyledChart>
+        <StyledCalendar>
+          <Calendar />
+          <ExplainCa>🟩：dot済み (クリックで確認)</ExplainCa>
+        </StyledCalendar>
+        <WeekStudyHours>
+          <StudyHours>今週の学習時間 / {week_hours}時間</StudyHours>
+          <StudyHours>前週の学習時間 / {lastweek_hours}時間</StudyHours>
+          <StudyHours>総学習時間 / {total_hours}時間</StudyHours>
+        </WeekStudyHours>
+      </form>
       <Footer />
-
-      {/* <div>Index</div>
-      <button style={{ width: "100px", height: "30px" }} onClick={logout}>
-        ログアウト
-      </button>
-      <div style={jobstyle}>
-        <div style={topjobstyle}>
-          <h4>Justice!!!!</h4> */}
-      <Link to="/mydots">
-        <button>MyDotsページ</button>
-      </Link>
-      {/* <Link to="/ranking"> */}
-      {/* <button>Rankingページ</button> */}
-      {/* </Link> */}
-      {/* </div> */}
-      {/* <div style={topjobstyle}>
-          <h4>goto</h4>
-          <Link to="/form">
-            <button>Formページ</button>
-          </Link>
-          <Link to="/ourdots">
-            <button>OurDotsページ</button>
-          </Link>
-        </div>
-        <div style={topjobstyle}>
-          <h4>ito</h4>
-          <Link to="/dot/:id.jsx">
-            <button>DotDetailページ</button>
-          </Link>
-          <Link to="/dot/:id/edit">
-            <button>Editページ</button>
-          </Link>
-        </div>
-        <div style={topjobstyle}>
-          <h4>iwaswa</h4>
-          <Link to="/home">
-            <button>Homeページ</button>
-          </Link>
-        </div>
-      </div> */}
     </React.Fragment>
   );
 }
